chore!: api keys are hashed in database, document Bitfield, change default port to 3651

+++ database/src/main/kotlin/org/noelware/charted/database/flags/OrganizationFlags.kt
@@ -20,8 +20,8 @@ package org.noelware.charted.database.flags
 import org.noelware.charted.common.Bitfield
 
 private val FLAGS = mapOf(
-    "PRIVATE" to (1 shl 0).toLong(), // Organization is private and only the members who are in the organization has access to its repositories.
-    "EXPERIMENTS" to (1 shl 1).toLong() // This organization has access to experimental features and can be enabled from the dashboard (/organizations/:name/settings#experimental)
+    "PRIVATE" to (1L shl 0), // Organization is private and only the members who are in the organization has access to its repositories.
+    "EXPERIMENTS" to (1L shl 1) // This organization has access to experimental features and can be enabled from the dashboard (/organizations/:name/settings#experimental)
 )
 
 class OrganizationFlags(originalBits: Long = 0): Bitfield(originalBits, FLAGS)
